Project Settings Spatial Audio Optimization
Had a breakthrough where I discovered that the number of Primary Rays in the Spatial Audio Project Settings is what actually causes the Diffraction path traces to move with less jitter (at the cost of CPU, which can be tuned with other settings). Will replicate settings on main Ambience branch.

NOTE: Unreal Project must be restarted for Project Settings changes to be applied.