Conclusion
Emitters are spawned below the floor at world origin. They move at the correct velocity and spawn at the same time as the hovercars, but need a properly shaped spline to match the hovercars + position offsets to work properly. Huge headache, will not be exploring this solution further. Thanks Jarrett for showing me the Vector Length conversion node (which makes the velocity useable when multiplying by time to get distance).