Fixed OnPlay Holograph FadeIn Click, and added Hovercar Wall
There is an unexplained click happening on play with the Holograph Fade In, so I added an imperceptible delay (0.01 seconds) and it fixed it. Likely an initialization problem since there's no OnPlay in the BP_Holo_Controller blueprint.

I also added a blocking wall with AkGeometry behind the building so that the Hovercars can't be heard through the walls of the Atrium Center Back